Caution! Only work in progress for discussion about the requirements listed in issues 6 (some also from 3).
I attemted to simplify the error reporting, but this was not fully tested.
Changes would be required in templates.go to print from the list.
In reviewing the golang documentaion it higlights the overhead involed
with array/slice append after the intialization length is utlized.

If the objective is to create tables for a number of diffrent database,
it might be easier to connect to the required instance and return
the DDL error to the user (for DDL syntax validation).

Also note that I am having more issues with github that golang, so
I am sure I am a few steps behind on these requirements.